Apply precedence rules to operator sections
This commit moves operator section desugaring from the beginning of the
desugaring pipeline to the rebracketing phase. This allows the compiler
to correctly detect whether an operator section is well-formed based on
the precedence rules in effect. (In the process, this commit also
refines the positioning of error messages associated with incorrect
anonymous arguments.)

Previously, `(_ * 4 + 1)` would desugar to `\x -> x * (4 + 1)`, even
though `*` has higher precedence than `+`. Conversely, `(3 * 2 + _)`
would not compile, even though `*` has higher precedence than `+`. These
bugs have now been fixed; `(_ * 4 + 1)` is an error, and `(3 * 2 + _)`
desugars to `\x -> 3 * 2 + x`.

If you have code that relied on the old behavior, add an extra pair of
parentheses around the expression in the section.